Moving low level peripheral interface and parsers into their own module
